Establishes common setup/teardown methods for tests
Tests now use the built-in tokio test runtime for async execution.

Adds setup/teardown methods to create and delete meetme services
accordingly.

Lastly, this flattens the directory structure of the tests project.
