+++ tests/src/common/find_service_setup.rs
@@ -3,13 +3,12 @@ use pubsub::find_service::Client;
 ///Wraps an external find service process and provides easy access to its functions
 pub struct FindService {
     _proc: tokio::process::Child,
-    _handle: tokio::runtime::Handle,
     client: pubsub::find_service::Client,
 }
 
 pub async fn retry_client(url: &'static str) -> pubsub::find_service::Client {
-    let retries = 10;
-    let mut retry = 0;
+    let retries: i32 = 10;
+    let mut retry: i32 = 0;
 
     loop {
         if retry >= retries {
@@ -33,34 +32,26 @@ pub async fn retry_client(url: &'static str) -> pubsub::find_service::Client {
 }
 
 impl FindService {
-    pub async fn new(_handle: tokio::runtime::Handle) -> FindService {
+    pub async fn new() -> FindService {
+        info!("Starting new meetme service");
         let path = "../target/debug/pubsub-meetup";
         let url = "http://127.0.0.1:8080/";
         let bind = "127.0.0.1:8080";
 
-        let _proc = _handle.enter(|| {
-            info!("Starting meetme service");
-            let ret =
-                tokio::process::Command::new(path)
-                    .arg("--bind")
-                    .arg(bind)
-                    .kill_on_drop(true)
-                    .spawn()
-                    .expect("Failed to start meetup");
+        let _proc = tokio::process::Command::new(path)
+            .arg("--bind")
+            .arg(bind)
+            .kill_on_drop(true)
+            .spawn()
+            .expect("Failed to start meetup");
 
-            info!("meetme service started");
-            //tokio::runtime::Handle::current().block_on(tokio::time::delay_for(std::time::Duration::from_secs(1)));
-            //info!("meetme service started");
-            ret
-        });
+        info!("meetme service started");
+        //tokio::runtime::Handle::current().block_on(tokio::time::delay_for(std::time::Duration::from_secs(1)));
+        //info!("meetme service started");
 
         let client = retry_client(url).await;
 
-        FindService {
-            _proc,
-            _handle,
-            client,
-        }
+        FindService { _proc, client }
     }
 
     pub fn client(&self) -> Client {
